% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stub.R
\name{stub}
\alias{stub}
\title{Build stubs out of functions}
\usage{
stub(function_to_stub)
}
\arguments{
\item{function_to_stub}{is the function that the user wants to make a stub out of}
}
\description{
See the vignette for usage details. You can access it by either visiting \url{https://cran.rstudio.com/web/packages/stubthat/vignettes/stubtthat.html} or by executing \code{vignette('stubthat')}.
}

